+++ LookupAnything/DataParser.cs
@@ -310,23 +310,27 @@ namespace Pathoschild.Stardew.LookupAnything
         /// <summary>Get the recipe ingredients.</summary>
         /// <param name="metadata">Provides metadata that's not available from the game data directly.</param>
         /// <param name="reflectionHelper">Simplifies access to private game code.</param>
-        public RecipeModel[] GetRecipes(Metadata metadata, IReflectionHelper reflectionHelper)
+        /// <param name="monitor">The monitor with which to log errors.</param>
+        public RecipeModel[] GetRecipes(Metadata metadata, IReflectionHelper reflectionHelper, IMonitor monitor)
         {
             List<RecipeModel> recipes = new List<RecipeModel>();
 
-            // cooking recipes
-            recipes.AddRange(
-                from entry in CraftingRecipe.cookingRecipes
-                let recipe = new CraftingRecipe(entry.Key, isCookingRecipe: true)
-                select new RecipeModel(recipe, reflectionHelper)
-            );
-
-            // crafting recipes
-            recipes.AddRange(
-                from entry in CraftingRecipe.craftingRecipes
-                let recipe = new CraftingRecipe(entry.Key, isCookingRecipe: false)
-                select new RecipeModel(recipe, reflectionHelper)
-            );
+            // cooking/crafting recipes
+            var craftingRecipes =
+                (from pair in CraftingRecipe.cookingRecipes select new { pair.Key, pair.Value, IsCookingRecipe = true })
+                .Concat(from pair in CraftingRecipe.craftingRecipes select new { pair.Key, pair.Value, IsCookingRecipe = false });
+            foreach (var entry in craftingRecipes)
+            {
+                try
+                {
+                    var recipe = new CraftingRecipe(entry.Key, entry.IsCookingRecipe);
+                    recipes.Add(new RecipeModel(recipe, reflectionHelper));
+                }
+                catch (Exception ex)
+                {
+                    monitor.Log($"Couldn't parse {(entry.IsCookingRecipe ? "cooking" : "crafting")} recipe '{entry.Key}' due to an invalid format.\nRecipe data: '{entry.Value}'\nError: {ex}", LogLevel.Warn);
+                }
+            }
 
             // machine recipes
             recipes.AddRange(